Moved deletion button to profile view, added confirmation view
