+++ src/routes/features/maintenance/[id]/+page.svelte
@@ -57,6 +57,9 @@
     <MenuItem icon="file-export" href="/features/maintenance/{profileId}/export">
         Export Profile
     </MenuItem>
+    <MenuItem href="/features/maintenance/{profileId}/delete">
+        Delete Profile
+    </MenuItem>
 </Menu>
 
 <style lang="scss">
